Architectural: extract functionality from godfile
- character_card_grid - stateless widget with most card manipulation logic and UI, pure-callback connection to home_page
